Added an option for not highlighting CamelCase links when the targeted pages are missing.
However, the link is still shown as missing if the CamelCase link
is enclosed in a description (e.g. `[wiki:DocumentThis To be documented]`
or even `[wiki:DocumentThis]`).

The behavior is made consistent for all the wiki rendering.

For enabling this feature, add the following to the [wiki:TracIni trac.ini]:
{{{
[wiki]
ignore_missing_pages = true
}}}


Should close #476.

''This changeset is dedicated to Pierre Brengard''


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@2135 af82e41b-90c4-0310-8c96-b1721e28e2e2