Reimplemented the quickjump feature by reusing the available TracLinks syntax.
First, this avoids the duplication of the regular expressions for tickets, wiki page names, etc.
Second, we gain the ability to quickjump to log ranges, revisions defined using `rXYZ` (#2881) and all the link resolvers that were not previously considered (attachment, query, log, etc.)

Likewise, this automatically extends the quickjump feature for TracLinks defined by the `IWikiSyntaxProvider` plugins (provided they return `Element`s objects with a `href` attribute, typically `html.A(href=...)`).

This required a small refactoring of the Formatter, a step in the direction of separating the parsing and the formatting phases in the Wiki engine.


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@3350 af82e41b-90c4-0310-8c96-b1721e28e2e2