Make the connection pool robust when the server uses a thread pool.
In a multi-threaded server, when a request is finished and `shutdown(tid)` is called, there might still exist some !PooledConnection objects for this `tid` which are not yet collected at this point. 

In the event the thread gets reused (*) and acquires again a connection from the pool, a new connection will be created and stored in the `_active[tid]` entry. The old !PooledConnection objects will therefore have a different cnx, and this is why the `assert` used to fail.

With this change, we simply acknowledge the fact that this can happen. Also, when this happens, there's no need to take special action as the "old" connection has already been closed during `shutdown(tid)`.

Related to #3504.

(*) even without a thread pool this could happen if the OS reuses the thread `tid` for a new thread. But in this case, the likelihood for having uncollected !PooledConnection with this `tid` still around seems to be low...

git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@3798 af82e41b-90c4-0310-8c96-b1721e28e2e2