TracQuery supports specifying lists of ids (fixes #4119)
The `ticket:` TracLinks has been extended so that one can specify lists (and even ranges) of tickets, e.g. `#12,23-34,35,36,50,40-42` (see #3011).

This works with the TicketQuery as well (#2604). However, in this context the tickets and ranges have to be separated by the "|" character, as the "," character is already used as an argument separator.

Example:
{{{
[[TicketQuery(group=type&id=12|23-36|40-42|50,format=table)]]
}}}

In addition, the "!" negation operator works also for such lists.


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@4239 af82e41b-90c4-0310-8c96-b1721e28e2e2