Remove remnants of old-style macros, as we couldn't support them anyway anymore.
The wiki-macros folder should be kept, though, at least for 0.11, as it now contains a README documenting the need for a conversion to the new-style macros. Likewise, the example macros (!HelloWorld and Timestamp) have been converted as single file plugins, and placed in the new `sample-plugins` folder.

The `sample-plugins` folder could be used for other, small, useful or otherwise illustrative plugin examples.


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@4587 af82e41b-90c4-0310-8c96-b1721e28e2e2