Transform into a warning any TracError that could happen during the early resync done for each request.
The inability to synchronize with the repository is a serious error which should be dealt with, but this shouldn't prevent the whole system (wiki, ticket system, etc.) to be used. To that end, the warning is the best compromise: the error is not masked, but the Trac instance is usable besides the versioncontrol subsystem.


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@5230 af82e41b-90c4-0310-8c96-b1721e28e2e2