Ticket #2048: add a `div` and a `span` built-in wiki processors.
The div macro should be used with a block and can be used for wrapping any wiki text (including contained blocks) within a <div> element. The `class` and `style` attributes of the <div> can be specified by listing keyword arguments after the div. The value of such an argument can be quoted using single or double quotes if it has to contain space characters (usually the case for the `style` attribute).

Example:
{{{
  {{{
  #!div class=important style="background: #ffaaaa"
  Note that:
    - use with care!
    - you have been warned
    - I won't repeat it
  }}}
}}}

The span macro should be used with the usual macro call style, like ![[span(Note: you have been warned, class=important, style=background:#ffbbbb)]]. Note that the arguments don't support quoting for now.

git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@6180 af82e41b-90c4-0310-8c96-b1721e28e2e2