Backout r6124, i.e. restore the r4886 fix, as this makes an essential difference '''on Linux'''. Without acquiring the db connection first, it's quite easy to trigger the timeout problem with simultaneous requests on that platform.
That variety of timeout is not of the "persistent" kind though, that's a different issue that will be dealt with separately.

Related to #4465.

Note that I've checked once again that this change makes absolutely no visible difference on Windows (i.e. with or without acquiring the db first, I can't trigger a timeout to happen).

git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@6260 af82e41b-90c4-0310-8c96-b1721e28e2e2