MultipleRepositorySupport: retrieve a reponame from a path
The path can contain the repository name (i.e. the ''key'' for retrieving the configuration data for that repository) and this should enable us to use a kind of "virtual" filesystem encompassing all the repositories.

In case there's a conflict between a path below one repository and a path below an other (for example with a path below the default repository), the longest repository name will have priority (therefore the default repository will never win).

API changes:
 - added `RepositoryManager.get_repository_by_name(path, authzname)`

git-svn-id: http://trac.edgewall.org/intertrac/log:/sandbox/multirepos@6449 af82e41b-90c4-0310-8c96-b1721e28e2e2