MultipleRepositorySupport: merge the repository index in the browser template.
This enables us to show at the same time the list of repositories and the top-level entries of the default repository (if any). All those "paths" are visible at the root, so it makes sense to display them all together.

Not to mention that the '''(default)''' repository link in the previous approach can't really be made to work.

Another benefit from this approach is that it's possible to browse in-place across all repositories (and in this case, browsing the default repository works... magic!).

git-svn-id: http://trac.edgewall.org/intertrac/log:/sandbox/multirepos@6464 af82e41b-90c4-0310-8c96-b1721e28e2e2