Changed the syntax for `[svn:externals]` TracIni section.
This is an incompatible change, but the former syntax was broken and unusable anyway. The !ConfigParser impose a `key <sep> value` syntax for entries in configuration sections, so that there's no way to have the `key` to contain either `:` or `=` characters, which was actually the case for all the externals URLs.

We tried to account for that in the code, by "repairing" the key/value pairs, but that workaround could only work when there were only ''one'' external URL (or actually, at most one for each protocol).

This is fixed now in a more general way by forcing the use of "dummy" unique keys, e.g.
{{{
[svn:externals]
1 = svn://server/repos1 http://trac/proj1/browser/$path?rev=$rev
2 = svn://server/repos2 http://trac/proj2/browser/$path?rev=$rev
}}}

git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@6579 af82e41b-90c4-0310-8c96-b1721e28e2e2