More exhaustive support for e-mail addresses.
The regexp used so far missed some unusual ones, like those containing quotes or double quotes. We're not going to complexify the regexp too much, so this is probably still not perfect.

Closes #6609.

This respects the constraints introduced by #4372 at the notification level, but doesn't yet inject the `[notification] admit_domains` in the Wiki regexp.


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@6676 af82e41b-90c4-0310-8c96-b1721e28e2e2