The !CachedRepository `next_rev` method didn't compare paths on directory boundaries, which could lead to subtle but very annoying issues.
For example, it was possible to "jump" from one branch to another one when following the next/prev change links on restricted changesets, provided those branch names had a common prefix. When displaying that other changeset on the wrong path, it appeared to be empty (No files), which by itself was correct but very confusing given the way you got there.


git-svn-id: http://trac.edgewall.org/intertrac/log:/branches/0.11-stable@7306 af82e41b-90c4-0310-8c96-b1721e28e2e2