0.11-stable: Fixed sorting by milestone in ticket query.
This changeset makes sorting of ticket query results by milestone consistent with the ordering of milestones in the roadmap:

 * First, completed milestones by completed date.
 * Then, due milestones by due date.
 * Finally, milestones without completed or due dates alphabetically.

One difference with the roadmap is that the alphabetical sort is not a "natural sort". So for milestones without completed or due dates, 0.10 will come before 0.9. There doesn't seem to be a way to do natural sorting in an SQL query.

The changeset also simplifies the management of ascending / descending sorting.

Closes #2209.


git-svn-id: http://trac.edgewall.org/intertrac/log:/branches/0.11-stable@7483 af82e41b-90c4-0310-8c96-b1721e28e2e2