MultiRepos: changed the way changesets are identified globally across repositories.
As I wanted to implement `SubversionChangeset.get_uid`, I figured it would be simpler to have the method on the level of the repository, as the trac.versioncontrol.cache layer doesn't know about backend specific changesets and would therefore have to delegate the call to the backend specific Repository class anyway.

Therefore `Changeset.get_uid` is now `Repository.get_changeset_uid` and I implemented the latter for `SubversionRepository`.


git-svn-id: http://trac.edgewall.org/intertrac/log:/sandbox/multirepos@8441 af82e41b-90c4-0310-8c96-b1721e28e2e2