Follow-up to the merge of the expanded unicode support in Config (r8469).
This broke the creation of query links in ticket view (#7562), as keyword arguments can't be unicode objects in Python 2.x. So we use the already existing feature of Href, which can be given a dict of parameters in its last positional argument.

Also checked other places that could have similar issues.

Completes #8276 on trunk.


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@8471 af82e41b-90c4-0310-8c96-b1721e28e2e2