psycopg2 and MySQLdb didn't handle correctly `execute()` arguments which were subclasses of `unicode`.
For psycopg2, we register explicit adapters.
For MySQLdb, we reuse the converter of the types.UnicodeType for the Markup class (Empty was OK since string_literal can convert it to `''`).

Fixes #8740 (Empty class, introduced in r8661) and #3501 (Markup class).

Thanks to harald.radi@nme.at for contributing a fix in #3501.


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@8669 af82e41b-90c4-0310-8c96-b1721e28e2e2