Add a simple syntax for defining anchors anywhere in wiki pages: {{{ [=#point1] }}}
The syntax is reminiscent of the syntax used for adding explicit ids to headings:
{{{
== Long title == #title
}}}

It's also very close to the syntax for the corresponding ''link'' to that anchor:
{{{
[#point1]
}}}

Optionally, a label can be given to the anchor:
{{{
[[=#point1 '''Point 1''']]
}}}

For more complex anchors (e.g. when a custom title is wanted), one can use the Span macro, e.g. `[[span(id=point1, class=wikianchor, title=Point 1, ^(1)^)]]`.

Closes #6413.


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@8697 af82e41b-90c4-0310-8c96-b1721e28e2e2