TracWiki: enhanced wiki editor.
In the default editing mode, the preview is now below the editor form.
This is the same user interface as for other popular wikis (MoinMoin,
Wikipedia, JspWiki).
Scrolling is minimized, as buttons are always close to the editor.

In addition, there's a new side-by-side editing mode, quite useful for
making review changes, as you don't have to jump back and forth between
the preview and the editor.

Closes #8721 and also addresses #8608 for wiki pages.


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@8888 af82e41b-90c4-0310-8c96-b1721e28e2e2