0.12.2dev: for rst rendering with docutils, don't store Trac env and context in the directive and role handlers.
This is not thread-safe. Instead, we register the "trac" and
"code-block" directives and roles only once and the handlers retrieve
the Trac environment and rendering context from the input we feed to
the docutils API.

As a bonus, Trac plugins can now use the same mechanism and register
their own directives and role handlers. For retrieving contextual
information, they can do the same way as we do here, look up the
Inliner's `trac` attribute. The attribute will be an `(env, context)`
pair and will be set if and only if Trac rendered the rst content. The
Inliner is available directly as the `inliner` argument passed to role
handlers and from the `.inliner` attribute of the `state` argument
passed to the directive handler.

git-svn-id: http://trac.edgewall.org/intertrac/log:/branches/0.12-stable@10458 af82e41b-90c4-0310-8c96-b1721e28e2e2