versioncontrol: refactor the `diff` module
While marking `hdf_diff` as deprecated, I noticed that
`_grouped_opcodes()` was introduced to compensate for the lack
of `SequenceMatcher.get_grouped_opcodes()` method in Python 2.2.

Now we can use it, but we can't throw away `_grouped_opcodes()`
completely either, as we need it in case the groups need to be
reformed if changes have been filtered out due to the `ignore_*`
options.

When no ignore options is set, we just return the opcodes directly
from the `SequenceMatcher`.

git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@10574 af82e41b-90c4-0310-8c96-b1721e28e2e2