web: cope with mod_wsgi collapsing multiple "/" in PATH_INFO.
This collapsing doesn't happen with SCRIPT_URL, hence the determination of SCRIPT_NAME from SCRIPT_URL and PATH_INFO needs to be fixed.

A less severe but similar issue can be observed with the "autoexpansion" feature, when we compare the browser `window.location.pathname` (multiple "/"s when browsing within the default repository) with the generated href ("/" collapsed).

Part of #10048.

git-svn-id: http://trac.edgewall.org/intertrac/log:/branches/0.12-stable@10609 af82e41b-90c4-0310-8c96-b1721e28e2e2