1.0.2dev: fix a subtle error in `Environment.get_version`.
During the adaptation to the 1.0 database API, we went from
`return row and ...` to `return rows and ...`. 
In case no rows was present, we returned `None` in the first case,
`[]` in the second, and these values don't behave the same way when
compared to ints.  Lesson learned: don't play tricks on the type
system, it'll hit back...

Patch by Ethan Jucovy <ethan.jucovy@gmail.com>, thanks!

Closes #11098.


git-svn-id: http://trac.edgewall.org/intertrac/log:/branches/1.0-stable@11740 af82e41b-90c4-0310-8c96-b1721e28e2e2