1.2dev: Render INSTALL and UPGRADE as ReST.
Renamed `INSTALL` to `INSTALL.rst` and `UPGRADE` to `UPGRADE.rst`.
Modified the default value of `[mimeviewer] mime_map_patterns` so
that `INSTALL.rst` isn't matched by the regular expression for the
`text/plain` mime type. `INSTALL.rst` and `UPGRADE.rst` should now
be rendered to HTML when docutils is installed, even when viewing
the files in a Git repository (or any other VCS that doesn't allow
the mimetype to be set as a repository property, like Subversion
does).

Refs #12141.


git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@14330 af82e41b-90c4-0310-8c96-b1721e28e2e2