1.0.10dev: add configuration file for AppVeyor.
This specifies the following set of build environments:
 - Python 2.7 x86 - SQLite (memory db)
 - Python 2.7 x64 - SQLite (file db)
 - Python 2.6 x86 - SQLite (file db)
 - Python 2.6 x64 - SQLite (memory db)
 - Python 2.7 (Conda) x86 - MySQL
 - Python 2.7 (Conda) x64 - PostgreSQL

The version of Python used is specified via the `PYTHONHOME`
environment variable, and the database backend is specified as usual
using the `TRAC_TEST_DB_URI` environment variable.

In addition, we no longer can assume that the `APPVEYOR_REPO_BRANCH`
environment variable will be 1.0-stable or trunk, as it could be the
name of any topic branch. We indicate the "base" branch (for the
package requirements) using the `SVN_BRANCH` environment variable (so
`1.0-stable` here).

The "script" parts (init, install, build_script, test_script) are as
minimal as it gets, they simply call the appropriate function loaded
from the `contrib/appveyor.ps1` script.


git-svn-id: http://trac.edgewall.org/intertrac/log:/branches/1.0-stable@14482 af82e41b-90c4-0310-8c96-b1721e28e2e2