1.0.10dev: introduce tearDown helper method `EnvironmentStup.reset_db_and_disk`
This ensures safe deletion of environment, even on Windows where it's
imperative to not retain an open file descriptor on a file that needs
to be deleted. As this is what happens with open connections on
file-based SQLite databases, we make sure we close the db connections
first.

Related to #12310.


git-svn-id: http://trac.edgewall.org/intertrac/log:/branches/1.0-stable@14493 af82e41b-90c4-0310-8c96-b1721e28e2e2