1.0.11dev: for AppVeyor, add Subversion bindings to the 32-bits environments
The bindings are fetched from the win32svn project once and then
stored in the dependencies.

There's no convenient way yet to add the bindings for 64-bits environments,
and the bindings are not available for Subversion 1.9.x, so it's not
perfect but better than nothing.

Note that the presence of a `svn` package for Conda is a bit misleading,
as it only contains the svn binaries, not the Python bindings.



git-svn-id: http://trac.edgewall.org/intertrac/log:/branches/1.0-stable@14611 af82e41b-90c4-0310-8c96-b1721e28e2e2