1.0.14dev: add release related tasks to the Makefile
This is the first half of the Trac release automation enhancements.

`make release` builds the packages relevant for each platform
(`release-exe` builds the .exe installers on Windows and `release-src`
the source packages on Unix).

`make upload version=1.x.y` then copies the packages on lynx, in the
user's `dist` folder over there, which has to exist.

The second half is in the `update-latest.sh` script on lynx, which
expects the packages to be in the `$SUDO_USER`'s `~/dist` folder, and
moves them to the appropriate places for distribution on t.e.o.


git-svn-id: http://trac.edgewall.org/intertrac/log:/branches/1.0-stable@15277 af82e41b-90c4-0310-8c96-b1721e28e2e2