(#12639): web and chrome API changes for Jinja2 vs. Genshi
This is were the choice to remain backward compatible with Genshi has
the most impact, as we can't actually distinguish between a Genshi
template and a Jinja2 template automatically.

Trying to load a Genshi template with the Jinja2 engine would have
impredictible results, i.e. in the worst case, it might even work and
not raise any error. And if it raises an error, it can be anything, so
we can't distinguish between a genuine error in a Jinja2 template and
the jinja engine loading a Genshi template.

Bottom line, it is better to make things explicit and have a way for
the plugin programmer to return either a Genshi template or a Jinja2
template. The way to "signal" that a template has to be rendered with
Genshi is of course to change nothing, i.e. keep the current
`IRequestHandler` and `IRequestFilter` APIs so that in the best case,
a plugin using Genshi templates doesn't have to be modified.

This forces us to change the API for signaling that a plugin (or Trac
itself for that matter) uses a Jinja2 template. For that, we can go
back to returning a simple `(template, data)` pair, like we had in the
ClearSilver days. By now, no one should be using those plugins
anymore, so there's no ambiguity possible.

All instances of `process_request` will have to be modified the
following way, from:
{{{#!py
    def process_request(req):
        """Generate my beautiful content with a Genshi template"""

	# ...

	return 'template.html', data, None
}}}

To:
{{{#!py
    def process_request(req):
        """Generate my beautiful content with a nice Jinja2 template"""

	# ...

	return 'template.html', data
}}}

The third component of the tuple (the content type) was most of the
time `None` anyway, so getting rid of it simplifies things.

In the few situations where the content type //had// to be specified
explicitly, this can now be done by passing a `metadata` dict as third
argument, with `'content_type'` as a key.

For example:
{{{#!py
            return 'revisionlog.txt', data, {'content_type': 'text/plain'}
}}}

The advantage is that we can pass more hints for the renderer in an
extensible way. For example, we can also pass the method by which the
engine should render the content (e.g. `{'method': 'text'}` which
means no auto-escaping).

To sum up, this are the old, backward compatible, return convention
for `process_request`:
 - `return (template, data, 'text/plain')` - `template` is the name of
   a Genshi template, the content type is `'text/plain'`
 - `return (template, data, None)` - `template` is the name of
   a Genshi template, the content type will default to `'text/html'`

This are the new return convention for `process_request`:
 - `return (template, data)` - `template` is the name of a Jinja2 template,
    the content type will default to `'text/html'`
 - `return (template, data, {'content_type': 'text/plain'})` -
   `template` is the name of a Jinja2 template, the content type is
   `'text/plain'`

The old API will be kept during the 1.3.x and 1.4.x transition period,
and removed, together with the dependency on Genshi, in the Trac 1.5.1
release.



git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@15443 af82e41b-90c4-0310-8c96-b1721e28e2e2