(#12639): final chrome API changes for Jinja2
A fix for the notification tests "degenerated" in some major API
clean-up on the Chrome component, hopefully for the better.

This is inspired by the previous web API change where the difference
between Genshi and Jinja2 templates was made by using "overloaded"
methods. Here, the differentiator is the way used to specify that a
text template should be generated instead of an XML/HTML one. This
choice for text vs. HTML has several implications for both template
engines, at different levels.

We'll focus on Jinja2 in the following description.

For enabling the XML/HTML auto-escaping feature of Jinja2, this used
to be done with a poor heuristic based on the template filename. This
was not satisfying but we couldn't do really better for lack of more
context when making the decision. Now I changed that to use 2
different `jinja2.Environment`, one with the feature and the other
without. The choice of which environment is made by using a `text`
boolean flag in the `Chrome.load_template` method. This contrasts with
Genshi where the specification for using the text mode came via a
`method` set to `'text'`. When `key` has a string value instead of a
boolean value, we detect that the legacy API is used and go the Genshi
way.

This style is propagated to the other methods as well.

The `text=True` mode has other effects:
 - return a plain `unicode` string instead of `Markup` in
   `Chrome.render_template_as_string`
 - don't sanitize the bytes generated in stream mode, in
   `Chrome.generate_template_stream`

The tests should all pass at this point.



git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@15445 af82e41b-90c4-0310-8c96-b1721e28e2e2