(#12640, #633) TracWiki restyling: enhance readability and visual impact.
The main change for the wiki is a default presentation of the page in
a central column of fixed width. Some elements won't fit in this column
(e.g. big tables, images) and having them overflow on the right side
wouldn't be nice, so we automatically "recenter" these elements to have
them overflow equally on both sides of this central column. If this is not
enough and reverting to the old "full width" behavior is desirable, one
can still do so by using an expander button on the upper right corner of
the page.

By default, the page is presented in narrow mode, but we'll turn that
into an UI preference.

Some other minor changes were done for the wiki:
 - the formatter no longer produces trailing `?` for links to missing
   wiki page, this is done in CSS now; the advantage is that when
   copy/pasting the question marks are left out
 - we use linear gradients to create h2 and h3 underline and the
   styling of wiki headings is applied via an explicit .section
   class
 - collapse margin for consecutive headings
 - restyle the wiki-toc
 - change highlight of links
 - restyle the tt and pre elements, and tweak the box shadows to
   make the box edges look less blurry
 - specify charset for the .css (sometimes the `content: ...`
   in UTF-8 was wrongly interpreted as latin1 in Chrome)
 - move the JavaScript of wiki_view.html in a file of its own, wiki.js

Also we start to give names to reused values in the .css files (using
comments), aiming at using Jinja2 templates for generating the CSS files
at a later point.

For consistency, some of the visual changes made in the wiki module
have been spread to other modules:
 - harmonize monospace in source browser with monospace in wiki.
 - increase font size for revision numbers
 - in browser, make dir entries bolder than file entries
 - the admin left-side menu styling is more in line with the rest
 - redo the styling of mainnav tabs to use a linear gradient instead
   of a pixmap, uppercase the tab name and fix "blur effect" bug (#11387)

We now also make systematic use of relative font sizes (#633) by using
em units, but we also keep percentages when they were already used.



git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@15523 af82e41b-90c4-0310-8c96-b1721e28e2e2