This change facilitates better ordering of search results.
The `ISearchSource` `get_search_results()` gets an optional sixth value
in the return tuple called `order` which is supposed to contain a
quadtuple which is to be composed of
 * rating for pagename/filename
 * rating for headings
 * rating for description/content
 * timestamp
The `order` parameter is then used to sort search results. The ratings
are evaluated in above order. In each case, a higher value will signify
a more relevant search result.